backend/requirements.txt
flask==3.1.0
openai==1.68.2
pillow==11.1.0
python-dotenv==1.1.0
pytest==8.3.5
requests==2.32.3
gunicorn==21.2.0 